fix: resolve many bugs

+++ app/Repositories/LikeRepository.php
@@ -10,24 +10,36 @@ class LikeRepository implements LikeRepositoryInterface
 {
     public function like($postId)
     {
+        $userId = auth()->id();
+
         $post = Post::find($postId);
         if (!$post) {
             throw ObjectExcpetions::InvalidPost();
         }
 
-        $post->likes()->create([
-            "user_id" => auth()->id(),
-        ]);
+        $alreadyLiked = $post->likes()->where('user_id', $userId)->exists();
+
+        if (!$alreadyLiked) {
+            $post->likes()->create([
+                "user_id" => auth()->id(),
+            ]);
+        }
     }
 
     public function unlike($postId)
     {
+        $userId = auth()->id();
+
         $post = Post::find($postId);
         if (!$post) {
             throw ObjectExcpetions::InvalidPost();
         }
 
-        $post->likes()->where("user_id", auth()->id())->delete();
+        $alreadyLiked = $post->likes()->where('user_id', $userId)->exists();
+
+        if ($alreadyLiked) {
+            $post->likes()->where("user_id", auth()->id())->delete();
+        }
     }
 
     public function getLikes($postId)
